+++ tools/libexec/perl/TTP.pm
@@ -862,7 +862,21 @@ sub run {
 	print STDERR Dumper( @ARGV ) if $ENV{TTP_DEBUG};
 	$ep = TTP::EP->new();
 	$ep->bootstrap();
-	TTP::RunnerCommand::runCommand( $ep );
+	my $command = TTP::RunnerCommand::runCommand( $ep );
+	return $command;
+}
+
+# -------------------------------------------------------------------------------------------------
+# Run by an external command to initialize a TTP context
+# Expects $0 be the full path name to the command script (this is the case in Windows+Strawberry)
+# and @ARGV the command-line arguments
+
+sub runExtern {
+	print STDERR Dumper( @ARGV ) if $ENV{TTP_DEBUG};
+	$ep = TTP::EP->new();
+	$ep->bootstrap();
+	my $command = TTP::RunnerExtern::runCommand( $ep );
+	return $command;
 }
 
 # -------------------------------------------------------------------------------------------------